Isolate gRPC service class definition
Isolate gRPC service class definition to make it easy
for writting unit tests.
